Fix some C overflows and dereferncies

+++ iop/smap_udpbd/src/ministack.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 
 
-#define IP_ADDR(a, b, c, d) ((a << 24) | (b << 16) | (c << 8) | d)
+#define IP_ADDR(a, b, c, d) (((uint8_t)(a) << 24) | ((uint8_t)(b) << 16) | ((uint8_t)(c) << 8) | (uint8_t)(d))
 
 static inline uint32_t htonl(uint32_t n)
 {
